consolidated reflection and base and removed DB2 reflector and dialect
