+++ sqlalchemy_ibmi/base.py
@@ -22,90 +22,101 @@
 import datetime, re
 from sqlalchemy import types as sa_types
 from sqlalchemy import schema as sa_schema
-from sqlalchemy import util
-from sqlalchemy.sql import compiler
-from sqlalchemy.sql import operators
-from sqlalchemy.engine import default
+from sqlalchemy import util, sql, Table, MetaData, Column
+from sqlalchemy.sql import compiler, operators
+from sqlalchemy.engine import default, reflection
 from sqlalchemy import __version__ as SA_Version
-from . import reflection as ibm_reflection
 
-from sqlalchemy.types import BLOB, CHAR, CLOB, DATE, DATETIME, INTEGER,\
-    SMALLINT, BIGINT, DECIMAL, NUMERIC, REAL, TIME, TIMESTAMP,\
+from sqlalchemy.types import BLOB, CHAR, CLOB, DATE, DATETIME, INTEGER, \
+    SMALLINT, BIGINT, DECIMAL, NUMERIC, REAL, TIME, TIMESTAMP, \
     VARCHAR, FLOAT
 
 SA_Version = [int(ver_token) for ver_token in SA_Version.split('.')[0:2]]
 
+
+class CoerceUnicode(sa_types.TypeDecorator):
+    def process_literal_param(self, value, dialect):
+        pass
+
+    impl = sa_types.Unicode
+
+    def process_bind_param(self, value, dialect):
+        if isinstance(value, str):
+            value = value
+        return value
+
+
 # as documented from:
 # http://publib.boulder.ibm.com/infocenter/db2luw/v9/index.jsp?topic=/com.ibm.db2.udb.doc/admin/r0001095.htm
 RESERVED_WORDS = set(
-   ['activate', 'disallow', 'locale', 'result', 'add', 'disconnect', 'localtime',
-    'result_set_locator', 'after', 'distinct', 'localtimestamp', 'return', 'alias',
-    'do', 'locator', 'returns', 'all', 'double', 'locators', 'revoke', 'allocate', 'drop',
-    'lock', 'right', 'allow', 'dssize', 'lockmax', 'rollback', 'alter', 'dynamic',
-    'locksize', 'routine', 'and', 'each', 'long', 'row', 'any', 'editproc', 'loop',
-    'row_number', 'as', 'else', 'maintained', 'rownumber', 'asensitive', 'elseif',
-    'materialized', 'rows', 'associate', 'enable', 'maxvalue', 'rowset', 'asutime',
-    'encoding', 'microsecond', 'rrn', 'at', 'encryption', 'microseconds', 'run',
-    'attributes', 'end', 'minute', 'savepoint', 'audit', 'end-exec', 'minutes', 'schema',
-    'authorization', 'ending', 'minvalue', 'scratchpad', 'aux', 'erase', 'mode', 'scroll',
-    'auxiliary', 'escape', 'modifies', 'search', 'before', 'every', 'month', 'second',
-    'begin', 'except', 'months', 'seconds', 'between', 'exception', 'new', 'secqty',
-    'binary', 'excluding', 'new_table', 'security', 'bufferpool', 'exclusive',
-    'nextval', 'select', 'by', 'execute', 'no', 'sensitive', 'cache', 'exists', 'nocache',
-    'sequence', 'call', 'exit', 'nocycle', 'session', 'called', 'explain', 'nodename',
-    'session_user', 'capture', 'external', 'nodenumber', 'set', 'cardinality',
-    'extract', 'nomaxvalue', 'signal', 'cascaded', 'fenced', 'nominvalue', 'simple',
-    'case', 'fetch', 'none', 'some', 'cast', 'fieldproc', 'noorder', 'source', 'ccsid',
-    'file', 'normalized', 'specific', 'char', 'final', 'not', 'sql', 'character', 'for',
-    'null', 'sqlid', 'check', 'foreign', 'nulls', 'stacked', 'close', 'free', 'numparts',
-    'standard', 'cluster', 'from', 'obid', 'start', 'collection', 'full', 'of', 'starting',
-    'collid', 'function', 'old', 'statement', 'column', 'general', 'old_table', 'static',
-    'comment', 'generated', 'on', 'stay', 'commit', 'get', 'open', 'stogroup', 'concat',
-    'global', 'optimization', 'stores', 'condition', 'go', 'optimize', 'style', 'connect',
-    'goto', 'option', 'substring', 'connection', 'grant', 'or', 'summary', 'constraint',
-    'graphic', 'order', 'synonym', 'contains', 'group', 'out', 'sysfun', 'continue',
-    'handler', 'outer', 'sysibm', 'count', 'hash', 'over', 'sysproc', 'count_big',
-    'hashed_value', 'overriding', 'system', 'create', 'having', 'package',
-    'system_user', 'cross', 'hint', 'padded', 'table', 'current', 'hold', 'pagesize',
-    'tablespace', 'current_date', 'hour', 'parameter', 'then', 'current_lc_ctype',
-    'hours', 'part', 'time', 'current_path', 'identity', 'partition', 'timestamp',
-    'current_schema', 'if', 'partitioned', 'to', 'current_server', 'immediate',
-    'partitioning', 'transaction', 'current_time', 'in', 'partitions', 'trigger',
-    'current_timestamp', 'including', 'password', 'trim', 'current_timezone',
-    'inclusive', 'path', 'type', 'current_user', 'increment', 'piecesize', 'undo',
-    'cursor', 'index', 'plan', 'union', 'cycle', 'indicator', 'position', 'unique', 'data',
-    'inherit', 'precision', 'until', 'database', 'inner', 'prepare', 'update',
-    'datapartitionname', 'inout', 'prevval', 'usage', 'datapartitionnum',
-    'insensitive', 'primary', 'user', 'date', 'insert', 'priqty', 'using', 'day',
-    'integrity', 'privileges', 'validproc', 'days', 'intersect', 'procedure', 'value',
-    'db2general', 'into', 'program', 'values', 'db2genrl', 'is', 'psid', 'variable',
-    'db2sql', 'isobid', 'query', 'variant', 'dbinfo', 'isolation', 'queryno', 'vcat',
-    'dbpartitionname', 'iterate', 'range', 'version', 'dbpartitionnum', 'jar', 'rank',
-    'view', 'deallocate', 'java', 'read', 'volatile', 'declare', 'join', 'reads', 'volumes',
-    'default', 'key', 'recovery', 'when', 'defaults', 'label', 'references', 'whenever',
-    'definition', 'language', 'referencing', 'where', 'delete', 'lateral', 'refresh',
-    'while', 'dense_rank', 'lc_ctype', 'release', 'with', 'denserank', 'leave', 'rename',
-    'without', 'describe', 'left', 'repeat', 'wlm', 'descriptor', 'like', 'reset', 'write',
-    'deterministic', 'linktype', 'resignal', 'xmlelement', 'diagnostics', 'local',
-    'restart', 'year', 'disable', 'localdate', 'restrict', 'years', '', 'abs', 'grouping',
-    'regr_intercept', 'are', 'int', 'regr_r2', 'array', 'integer', 'regr_slope',
-    'asymmetric', 'intersection', 'regr_sxx', 'atomic', 'interval', 'regr_sxy', 'avg',
-    'large', 'regr_syy', 'bigint', 'leading', 'rollup', 'blob', 'ln', 'scope', 'boolean',
-    'lower', 'similar', 'both', 'match', 'smallint', 'ceil', 'max', 'specifictype',
-    'ceiling', 'member', 'sqlexception', 'char_length', 'merge', 'sqlstate',
-    'character_length', 'method', 'sqlwarning', 'clob', 'min', 'sqrt', 'coalesce', 'mod',
-    'stddev_pop', 'collate', 'module', 'stddev_samp', 'collect', 'multiset',
-    'submultiset', 'convert', 'national', 'sum', 'corr', 'natural', 'symmetric',
-    'corresponding', 'nchar', 'tablesample', 'covar_pop', 'nclob', 'timezone_hour',
-    'covar_samp', 'normalize', 'timezone_minute', 'cube', 'nullif', 'trailing',
-    'cume_dist', 'numeric', 'translate', 'current_default_transform_group',
-    'octet_length', 'translation', 'current_role', 'only', 'treat',
-    'current_transform_group_for_type', 'overlaps', 'true', 'dec', 'overlay',
-    'uescape', 'decimal', 'percent_rank', 'unknown', 'deref', 'percentile_cont',
-    'unnest', 'element', 'percentile_disc', 'upper', 'exec', 'power', 'var_pop', 'exp',
-    'real', 'var_samp', 'false', 'recursive', 'varchar', 'filter', 'ref', 'varying',
-    'float', 'regr_avgx', 'width_bucket', 'floor', 'regr_avgy', 'window', 'fusion',
-    'regr_count', 'within', 'asc'])
+    ['activate', 'disallow', 'locale', 'result', 'add', 'disconnect', 'localtime',
+     'result_set_locator', 'after', 'distinct', 'localtimestamp', 'return', 'alias',
+     'do', 'locator', 'returns', 'all', 'double', 'locators', 'revoke', 'allocate', 'drop',
+     'lock', 'right', 'allow', 'dssize', 'lockmax', 'rollback', 'alter', 'dynamic',
+     'locksize', 'routine', 'and', 'each', 'long', 'row', 'any', 'editproc', 'loop',
+     'row_number', 'as', 'else', 'maintained', 'rownumber', 'asensitive', 'elseif',
+     'materialized', 'rows', 'associate', 'enable', 'maxvalue', 'rowset', 'asutime',
+     'encoding', 'microsecond', 'rrn', 'at', 'encryption', 'microseconds', 'run',
+     'attributes', 'end', 'minute', 'savepoint', 'audit', 'end-exec', 'minutes', 'schema',
+     'authorization', 'ending', 'minvalue', 'scratchpad', 'aux', 'erase', 'mode', 'scroll',
+     'auxiliary', 'escape', 'modifies', 'search', 'before', 'every', 'month', 'second',
+     'begin', 'except', 'months', 'seconds', 'between', 'exception', 'new', 'secqty',
+     'binary', 'excluding', 'new_table', 'security', 'bufferpool', 'exclusive',
+     'nextval', 'select', 'by', 'execute', 'no', 'sensitive', 'cache', 'exists', 'nocache',
+     'sequence', 'call', 'exit', 'nocycle', 'session', 'called', 'explain', 'nodename',
+     'session_user', 'capture', 'external', 'nodenumber', 'set', 'cardinality',
+     'extract', 'nomaxvalue', 'signal', 'cascaded', 'fenced', 'nominvalue', 'simple',
+     'case', 'fetch', 'none', 'some', 'cast', 'fieldproc', 'noorder', 'source', 'ccsid',
+     'file', 'normalized', 'specific', 'char', 'final', 'not', 'sql', 'character', 'for',
+     'null', 'sqlid', 'check', 'foreign', 'nulls', 'stacked', 'close', 'free', 'numparts',
+     'standard', 'cluster', 'from', 'obid', 'start', 'collection', 'full', 'of', 'starting',
+     'collid', 'function', 'old', 'statement', 'column', 'general', 'old_table', 'static',
+     'comment', 'generated', 'on', 'stay', 'commit', 'get', 'open', 'stogroup', 'concat',
+     'global', 'optimization', 'stores', 'condition', 'go', 'optimize', 'style', 'connect',
+     'goto', 'option', 'substring', 'connection', 'grant', 'or', 'summary', 'constraint',
+     'graphic', 'order', 'synonym', 'contains', 'group', 'out', 'sysfun', 'continue',
+     'handler', 'outer', 'sysibm', 'count', 'hash', 'over', 'sysproc', 'count_big',
+     'hashed_value', 'overriding', 'system', 'create', 'having', 'package',
+     'system_user', 'cross', 'hint', 'padded', 'table', 'current', 'hold', 'pagesize',
+     'tablespace', 'current_date', 'hour', 'parameter', 'then', 'current_lc_ctype',
+     'hours', 'part', 'time', 'current_path', 'identity', 'partition', 'timestamp',
+     'current_schema', 'if', 'partitioned', 'to', 'current_server', 'immediate',
+     'partitioning', 'transaction', 'current_time', 'in', 'partitions', 'trigger',
+     'current_timestamp', 'including', 'password', 'trim', 'current_timezone',
+     'inclusive', 'path', 'type', 'current_user', 'increment', 'piecesize', 'undo',
+     'cursor', 'index', 'plan', 'union', 'cycle', 'indicator', 'position', 'unique', 'data',
+     'inherit', 'precision', 'until', 'database', 'inner', 'prepare', 'update',
+     'datapartitionname', 'inout', 'prevval', 'usage', 'datapartitionnum',
+     'insensitive', 'primary', 'user', 'date', 'insert', 'priqty', 'using', 'day',
+     'integrity', 'privileges', 'validproc', 'days', 'intersect', 'procedure', 'value',
+     'db2general', 'into', 'program', 'values', 'db2genrl', 'is', 'psid', 'variable',
+     'db2sql', 'isobid', 'query', 'variant', 'dbinfo', 'isolation', 'queryno', 'vcat',
+     'dbpartitionname', 'iterate', 'range', 'version', 'dbpartitionnum', 'jar', 'rank',
+     'view', 'deallocate', 'java', 'read', 'volatile', 'declare', 'join', 'reads', 'volumes',
+     'default', 'key', 'recovery', 'when', 'defaults', 'label', 'references', 'whenever',
+     'definition', 'language', 'referencing', 'where', 'delete', 'lateral', 'refresh',
+     'while', 'dense_rank', 'lc_ctype', 'release', 'with', 'denserank', 'leave', 'rename',
+     'without', 'describe', 'left', 'repeat', 'wlm', 'descriptor', 'like', 'reset', 'write',
+     'deterministic', 'linktype', 'resignal', 'xmlelement', 'diagnostics', 'local',
+     'restart', 'year', 'disable', 'localdate', 'restrict', 'years', '', 'abs', 'grouping',
+     'regr_intercept', 'are', 'int', 'regr_r2', 'array', 'integer', 'regr_slope',
+     'asymmetric', 'intersection', 'regr_sxx', 'atomic', 'interval', 'regr_sxy', 'avg',
+     'large', 'regr_syy', 'bigint', 'leading', 'rollup', 'blob', 'ln', 'scope', 'boolean',
+     'lower', 'similar', 'both', 'match', 'smallint', 'ceil', 'max', 'specifictype',
+     'ceiling', 'member', 'sqlexception', 'char_length', 'merge', 'sqlstate',
+     'character_length', 'method', 'sqlwarning', 'clob', 'min', 'sqrt', 'coalesce', 'mod',
+     'stddev_pop', 'collate', 'module', 'stddev_samp', 'collect', 'multiset',
+     'submultiset', 'convert', 'national', 'sum', 'corr', 'natural', 'symmetric',
+     'corresponding', 'nchar', 'tablesample', 'covar_pop', 'nclob', 'timezone_hour',
+     'covar_samp', 'normalize', 'timezone_minute', 'cube', 'nullif', 'trailing',
+     'cume_dist', 'numeric', 'translate', 'current_default_transform_group',
+     'octet_length', 'translation', 'current_role', 'only', 'treat',
+     'current_transform_group_for_type', 'overlaps', 'true', 'dec', 'overlay',
+     'uescape', 'decimal', 'percent_rank', 'unknown', 'deref', 'percentile_cont',
+     'unnest', 'element', 'percentile_disc', 'upper', 'exec', 'power', 'var_pop', 'exp',
+     'real', 'var_samp', 'false', 'recursive', 'varchar', 'filter', 'ref', 'varying',
+     'float', 'regr_avgx', 'width_bucket', 'floor', 'regr_avgy', 'window', 'fusion',
+     'regr_count', 'within', 'asc'])
 
 
 class _IBM_Boolean(sa_types.Boolean):
@@ -116,6 +127,7 @@ class _IBM_Boolean(sa_types.Boolean):
                 return None
             else:
                 return bool(value)
+
         return process
 
     def bind_processor(self, dialect):
@@ -126,7 +138,9 @@ class _IBM_Boolean(sa_types.Boolean):
                 return '1'
             else:
                 return '0'
+
         return process
+
 
 class _IBM_Date(sa_types.Date):
 
@@ -137,6 +151,7 @@ class _IBM_Date(sa_types.Date):
             if isinstance(value, datetime.datetime):
                 value = datetime.date(value.year, value.month, value.day)
             return value
+
         return process
 
     def bind_processor(self, dialect):
@@ -146,19 +161,25 @@ class _IBM_Date(sa_types.Date):
             if isinstance(value, datetime.datetime):
                 value = datetime.date(value.year, value.month, value.day)
             return str(value)
+
         return process
+
 
 class DOUBLE(sa_types.Numeric):
     __visit_name__ = 'DOUBLE'
 
+
 class LONGVARCHAR(sa_types.VARCHAR):
     __visit_name_ = 'LONGVARCHAR'
+
 
 class DBCLOB(sa_types.CLOB):
     __visit_name__ = "DBCLOB"
 
+
 class GRAPHIC(sa_types.CHAR):
     __visit_name__ = "GRAPHIC"
+
 
 class VARGRAPHIC(sa_types.Unicode):
     __visit_name__ = "VARGRAPHIC"
@@ -167,14 +188,16 @@ class VARGRAPHIC(sa_types.Unicode):
 class LONGVARGRAPHIC(sa_types.UnicodeText):
     __visit_name__ = "LONGVARGRAPHIC"
 
+
 class XML(sa_types.Text):
     __visit_name__ = "XML"
+
 
 colspecs = {
     sa_types.Boolean: _IBM_Boolean,
     sa_types.Date: _IBM_Date
-# really ?
-#    sa_types.Unicode: DB2VARGRAPHIC
+    # really ?
+    #    sa_types.Unicode: DB2VARGRAPHIC
 }
 
 ischema_names = {
@@ -206,7 +229,6 @@ ischema_names = {
 
 class DB2TypeCompiler(compiler.GenericTypeCompiler):
 
-
     def visit_TIMESTAMP(self, type_):
         return "TIMESTAMP"
 
@@ -230,7 +252,7 @@ class DB2TypeCompiler(compiler.GenericTypeCompiler):
 
     def visit_FLOAT(self, type_):
         return "FLOAT" if type_.precision is None else \
-                "FLOAT(%(precision)s)" % {'precision': type_.precision}
+            "FLOAT(%(precision)s)" % {'precision': type_.precision}
 
     def visit_DOUBLE(self, type_):
         return "DOUBLE"
@@ -243,11 +265,11 @@ class DB2TypeCompiler(compiler.GenericTypeCompiler):
 
     def visit_BLOB(self, type_):
         return "BLOB(1M)" if type_.length in (None, 0) else \
-                "BLOB(%(length)s)" % {'length': type_.length}
+            "BLOB(%(length)s)" % {'length': type_.length}
 
     def visit_DBCLOB(self, type_):
         return "DBCLOB(1M)" if type_.length in (None, 0) else \
-                "DBCLOB(%(length)s)" % {'length': type_.length}
+            "DBCLOB(%(length)s)" % {'length': type_.length}
 
     def visit_VARCHAR(self, type_):
         return "VARCHAR(%(length)s)" % {'length': type_.length}
@@ -263,11 +285,11 @@ class DB2TypeCompiler(compiler.GenericTypeCompiler):
 
     def visit_CHAR(self, type_):
         return "CHAR" if type_.length in (None, 0) else \
-                "CHAR(%(length)s)" % {'length': type_.length}
+            "CHAR(%(length)s)" % {'length': type_.length}
 
     def visit_GRAPHIC(self, type_):
         return "GRAPHIC" if type_.length in (None, 0) else \
-                "GRAPHIC(%(length)s)" % {'length': type_.length}
+            "GRAPHIC(%(length)s)" % {'length': type_.length}
 
     def visit_DECIMAL(self, type_):
         if not type_.precision:
@@ -276,8 +298,7 @@ class DB2TypeCompiler(compiler.GenericTypeCompiler):
             return "DECIMAL(%(precision)s, 0)" % {'precision': type_.precision}
         else:
             return "DECIMAL(%(precision)s, %(scale)s)" % {
-                            'precision': type_.precision, 'scale': type_.scale}
-
+                'precision': type_.precision, 'scale': type_.scale}
 
     def visit_numeric(self, type_):
         return self.visit_DECIMAL(type_)
@@ -317,17 +338,16 @@ class DB2TypeCompiler(compiler.GenericTypeCompiler):
 
 
 class DB2Compiler(compiler.SQLCompiler):
-
     if SA_Version < [0, 9]:
         def visit_false(self, expr, **kw):
             return '0'
-            
+
         def visit_true(self, expr, **kw):
             return '1'
-            
+
     def get_cte_preamble(self, recursive):
         return "WITH"
-        
+
     def visit_now_func(self, fn, **kw):
         return "CURRENT_TIMESTAMP"
 
@@ -338,12 +358,12 @@ class DB2Compiler(compiler.SQLCompiler):
             return ' WITH RS USE AND KEEP SHARE LOCKS'
         else:
             return ''
-			
+
     def visit_mod_binary(self, binary, operator, **kw):
         return "mod(%s, %s)" % (self.process(binary.left),
-                                                self.process(binary.right))
+                                self.process(binary.right))
 
-    def limit_clause(self, select,**kwargs):
+    def limit_clause(self, select, **kwargs):
         if (select._limit is not None) and (select._offset is None):
             return " FETCH FIRST %s ROWS ONLY" % select._limit
         else:
@@ -356,7 +376,7 @@ class DB2Compiler(compiler.SQLCompiler):
             __rownum = 'Z.__ROWNUM'
             sql_split = re.split("[\s+]FROM ", sql_ori, 1)
             sql_sec = ""
-            sql_sec = " \nFROM %s " % ( sql_split[1] )
+            sql_sec = " \nFROM %s " % (sql_split[1])
 
             dummyVal = "Z.__db2_"
             sql_pri = ""
@@ -365,45 +385,45 @@ class DB2Compiler(compiler.SQLCompiler):
             if select._distinct:
                 sql_sel = "SELECT DISTINCT "
 
-            sql_select_token = sql_split[0].split( "," )
+            sql_select_token = sql_split[0].split(",")
             i = 0
-            while ( i < len( sql_select_token ) ):
-                if sql_select_token[i].count( "TIMESTAMP(DATE(SUBSTR(CHAR(" ) == 1:
-                    sql_sel = "%s \"%s%d\"," % ( sql_sel, dummyVal, i + 1 )
+            while (i < len(sql_select_token)):
+                if sql_select_token[i].count("TIMESTAMP(DATE(SUBSTR(CHAR(") == 1:
+                    sql_sel = "%s \"%s%d\"," % (sql_sel, dummyVal, i + 1)
                     sql_pri = '%s %s,%s,%s,%s AS "%s%d",' % (
-                                    sql_pri,
-                                    sql_select_token[i],
-                                    sql_select_token[i + 1],
-                                    sql_select_token[i + 2],
-                                    sql_select_token[i + 3],
-                                    dummyVal, i + 1 )
+                        sql_pri,
+                        sql_select_token[i],
+                        sql_select_token[i + 1],
+                        sql_select_token[i + 2],
+                        sql_select_token[i + 3],
+                        dummyVal, i + 1)
                     i = i + 4
                     continue
 
-                if sql_select_token[i].count( " AS " ) == 1:
-                    temp_col_alias = sql_select_token[i].split( " AS " )
-                    sql_pri = '%s %s,' % ( sql_pri, sql_select_token[i] )
-                    sql_sel = "%s %s," % ( sql_sel, temp_col_alias[1] )
+                if sql_select_token[i].count(" AS ") == 1:
+                    temp_col_alias = sql_select_token[i].split(" AS ")
+                    sql_pri = '%s %s,' % (sql_pri, sql_select_token[i])
+                    sql_sel = "%s %s," % (sql_sel, temp_col_alias[1])
                     i = i + 1
                     continue
 
-                sql_pri = '%s %s AS "%s%d",' % ( sql_pri, sql_select_token[i], dummyVal, i + 1 )
-                sql_sel = "%s \"%s%d\"," % ( sql_sel, dummyVal, i + 1 )
+                sql_pri = '%s %s AS "%s%d",' % (sql_pri, sql_select_token[i], dummyVal, i + 1)
+                sql_sel = "%s \"%s%d\"," % (sql_sel, dummyVal, i + 1)
                 i = i + 1
 
-            sql_pri = sql_pri[:len( sql_pri ) - 1]
-            sql_pri = "%s%s" % ( sql_pri, sql_sec )
-            sql_sel = sql_sel[:len( sql_sel ) - 1]
-            sql = '%s, ( ROW_NUMBER() OVER() ) AS "%s" FROM ( %s ) AS M' % ( sql_sel, __rownum, sql_pri )
-            sql = '%s FROM ( %s ) Z WHERE' % ( sql_sel, sql )
+            sql_pri = sql_pri[:len(sql_pri) - 1]
+            sql_pri = "%s%s" % (sql_pri, sql_sec)
+            sql_sel = sql_sel[:len(sql_sel) - 1]
+            sql = '%s, ( ROW_NUMBER() OVER() ) AS "%s" FROM ( %s ) AS M' % (sql_sel, __rownum, sql_pri)
+            sql = '%s FROM ( %s ) Z WHERE' % (sql_sel, sql)
 
             if offset != 0:
-                sql = '%s "%s" > %d' % ( sql, __rownum, offset )
+                sql = '%s "%s" > %d' % (sql, __rownum, offset)
             if offset != 0 and limit is not None:
-                sql = '%s AND ' % ( sql )
+                sql = '%s AND ' % (sql)
             if limit is not None:
-                sql = '%s "%s" <= %d' % ( sql, __rownum, offset + limit )
-            return "( %s )" % ( sql, )
+                sql = '%s "%s" <= %d' % (sql, __rownum, offset + limit)
+            return "( %s )" % (sql,)
         else:
             return sql_ori
 
@@ -412,7 +432,7 @@ class DB2Compiler(compiler.SQLCompiler):
 
     def default_from(self):
         # DB2 uses SYSIBM.SYSDUMMY1 table for row count
-        return  " FROM SYSIBM.SYSDUMMY1"
+        return " FROM SYSIBM.SYSDUMMY1"
 
     def visit_function(self, func, result_map=None, **kwargs):
         if func.name.upper() == "AVG":
@@ -421,12 +441,12 @@ class DB2Compiler(compiler.SQLCompiler):
             return "CHAR_LENGTH(%s, %s)" % (self.function_argspec(func, **kwargs), 'OCTETS')
         else:
             return compiler.SQLCompiler.visit_function(self, func, **kwargs)
+
     # TODO: this is wrong but need to know what DB2 is expecting here
     #    if func.name.upper() == "LENGTH":
     #        return "LENGTH('%s')" % func.compile().params[func.name + '_1']
     #    else:
     #        return compiler.SQLCompiler.visit_function(self, func, **kwargs)
-
 
     def visit_cast(self, cast, **kw):
         type_ = cast.typeclause.type
@@ -435,13 +455,13 @@ class DB2Compiler(compiler.SQLCompiler):
         # other types, I was able to CAST against VARCHAR
         # for example
         if isinstance(type_, (
-                    sa_types.DateTime, sa_types.Date, sa_types.Time,
-                    sa_types.DECIMAL,sa_types.String)):
+                sa_types.DateTime, sa_types.Date, sa_types.Time,
+                sa_types.DECIMAL, sa_types.String)):
             return super(DB2Compiler, self).visit_cast(cast, **kw)
         else:
             return self.process(cast.clause)
 
-    def get_select_precolumns(self, select,**kwargs):
+    def get_select_precolumns(self, select, **kwargs):
         if isinstance(select._distinct, str):
             return select._distinct.upper() + " "
         elif select._distinct:
@@ -460,31 +480,32 @@ class DB2Compiler(compiler.SQLCompiler):
              self.process(join.onclause, **kwargs)))
 
     def visit_savepoint(self, savepoint_stmt):
-        return "SAVEPOINT %(sid)s ON ROLLBACK RETAIN CURSORS" % {'sid':self.preparer.format_savepoint(savepoint_stmt)}
+        return "SAVEPOINT %(sid)s ON ROLLBACK RETAIN CURSORS" % {'sid': self.preparer.format_savepoint(savepoint_stmt)}
 
     def visit_rollback_to_savepoint(self, savepoint_stmt):
-        return 'ROLLBACK TO SAVEPOINT %(sid)s'% {'sid':self.preparer.format_savepoint(savepoint_stmt)}
+        return 'ROLLBACK TO SAVEPOINT %(sid)s' % {'sid': self.preparer.format_savepoint(savepoint_stmt)}
 
     def visit_release_savepoint(self, savepoint_stmt):
-        return 'RELEASE TO SAVEPOINT %(sid)s'% {'sid':self.preparer.format_savepoint(savepoint_stmt)}
-    
+        return 'RELEASE TO SAVEPOINT %(sid)s' % {'sid': self.preparer.format_savepoint(savepoint_stmt)}
+
     def visit_unary(self, unary, **kw):
-        if (unary.operator == operators.exists)  and kw.get('within_columns_clause', False):
+        if (unary.operator == operators.exists) and kw.get('within_columns_clause', False):
             usql = super(DB2Compiler, self).visit_unary(unary, **kw)
             usql = "CASE WHEN " + usql + " THEN 1 ELSE 0 END"
             return usql
         else:
             return super(DB2Compiler, self).visit_unary(unary, **kw)
 
+
 class DB2DDLCompiler(compiler.DDLCompiler):
-    
+
     def get_server_version_info(self, dialect):
         """Returns the DB2 server major and minor version as a list of ints."""
         if hasattr(dialect, 'dbms_ver'):
             return [int(ver_token) for ver_token in dialect.dbms_ver.split('.')[0:2]]
         else:
             return []
-    
+
     def _is_nullable_unique_constraint_supported(self, dialect):
         """Checks to see if the DB2 version is at least 10.5.
         This is needed for checking if unique constraints with null columns are supported.
@@ -492,15 +513,14 @@ class DB2DDLCompiler(compiler.DDLCompiler):
 
         dbms_name = getattr(dialect, 'dbms_name', None)
         if hasattr(dialect, 'dbms_name'):
-           if dbms_name != None and (dbms_name.find('DB2/') != -1):
+            if dbms_name != None and (dbms_name.find('DB2/') != -1):
                 return self.get_server_version_info(dialect) >= [10, 5]
         else:
             return False
 
     def get_column_specification(self, column, **kw):
         col_spec = [self.preparer.format_column(column)]
-        col_spec.append(self.dialect.type_compiler.process(column.type,type_expression=column))
-
+        col_spec.append(self.dialect.type_compiler.process(column.type, type_expression=column))
 
         # column-options: "NOT NULL"
         if not column.nullable or column.primary_key:
@@ -551,14 +571,14 @@ class DB2DDLCompiler(compiler.DDLCompiler):
         else:
             qual = ""
             const = self.preparer.format_constraint(constraint)
-            
+
         if hasattr(constraint, 'uConstraint_as_index') and constraint.uConstraint_as_index:
             return "DROP %s%s" % \
-                                (qual, const)
+                   (qual, const)
         return "ALTER TABLE %s DROP %s%s" % \
-                                (self.preparer.format_table(constraint.table),
-                                qual, const)
-                                
+               (self.preparer.format_table(constraint.table),
+                qual, const)
+
     def create_table_constraints(self, table, **kw):
         if self._is_nullable_unique_constraint_supported(self.dialect):
             for constraint in table._sorted_constraints:
@@ -570,20 +590,21 @@ class DB2DDLCompiler(compiler.DDLCompiler):
                             break
                     if getattr(constraint, 'uConstraint_as_index', None):
                         if not constraint.name:
-                            index_name = "%s_%s_%s" % ('ukey', self.preparer.format_table(constraint.table), '_'.join(column.name for column in constraint))
+                            index_name = "%s_%s_%s" % ('ukey', self.preparer.format_table(constraint.table),
+                                                       '_'.join(column.name for column in constraint))
                         else:
                             index_name = constraint.name
                         index = sa_schema.Index(index_name, *(column for column in constraint))
                         index.unique = True
                         index.uConstraint_as_index = True
-        result = super( DB2DDLCompiler, self ).create_table_constraints(table, **kw)
+        result = super(DB2DDLCompiler, self).create_table_constraints(table, **kw)
         return result
-    
+
     def visit_create_index(self, create, include_schema=True, include_table_schema=True):
         if SA_Version < [0, 8]:
-            sql = super( DB2DDLCompiler, self ).visit_create_index(create)
+            sql = super(DB2DDLCompiler, self).visit_create_index(create)
         else:
-            sql = super( DB2DDLCompiler, self ).visit_create_index(create, include_schema, include_table_schema)
+            sql = super(DB2DDLCompiler, self).visit_create_index(create, include_schema, include_table_schema)
         if getattr(create.element, 'uConstraint_as_index', None):
             sql += ' EXCLUDE NULL KEYS'
         return sql
@@ -597,21 +618,23 @@ class DB2DDLCompiler(compiler.DDLCompiler):
                         break
                 if getattr(create.element, 'uConstraint_as_index', None):
                     if not create.element.name:
-                        index_name = "%s_%s_%s" % ('uk_index', self.preparer.format_table(create.element.table), '_'.join(column.name for column in create.element))
+                        index_name = "%s_%s_%s" % ('uk_index', self.preparer.format_table(create.element.table),
+                                                   '_'.join(column.name for column in create.element))
                     else:
                         index_name = create.element.name
                     index = sa_schema.Index(index_name, *(column for column in create.element))
                     index.unique = True
                     index.uConstraint_as_index = True
-                    sql = self.visit_create_index(sa_schema.CreateIndex(index)) 
+                    sql = self.visit_create_index(sa_schema.CreateIndex(index))
                     return sql
-        sql = super( DB2DDLCompiler, self ).visit_add_constraint(create)
+        sql = super(DB2DDLCompiler, self).visit_add_constraint(create)
         return sql
-    
-class DB2IdentifierPreparer(compiler.IdentifierPreparer):
 
+
+class DB2IdentifierPreparer(compiler.IdentifierPreparer):
     reserved_words = RESERVED_WORDS
     illegal_initial_characters = set(range(0, 10)).union(["_", "$"])
+
 
 class _SelectLastRowIDMixin(object):
     _select_lastrowid = False
@@ -627,29 +650,28 @@ class _SelectLastRowIDMixin(object):
             insert_has_sequence = seq_column is not None
 
             self._select_lastrowid = insert_has_sequence and \
-                                        not self.compiled.returning and \
-                                        not self.compiled.inline
+                                     not self.compiled.returning and \
+                                     not self.compiled.inline
 
     def post_exec(self):
         conn = self.root_connection
         if self._select_lastrowid:
             conn._cursor_execute(self.cursor,
-                    "SELECT IDENTITY_VAL_LOCAL() FROM SYSIBM.SYSDUMMY1",
-                    (), self)
+                                 "SELECT IDENTITY_VAL_LOCAL() FROM SYSIBM.SYSDUMMY1",
+                                 (), self)
             row = self.cursor.fetchall()[0]
             if row[0] is not None:
                 self._lastrowid = int(row[0])
 
-				
-class DB2ExecutionContext(_SelectLastRowIDMixin,default.DefaultExecutionContext):
+
+class DB2ExecutionContext(_SelectLastRowIDMixin, default.DefaultExecutionContext):
     def fire_sequence(self, seq, type_):
         return self._execute_scalar("SELECT NEXTVAL FOR " +
-                    self.dialect.identifier_preparer.format_sequence(seq) +
-                    " FROM SYSIBM.SYSDUMMY1", type_)
+                                    self.dialect.identifier_preparer.format_sequence(seq) +
+                                    " FROM SYSIBM.SYSDUMMY1", type_)
 
 
 class DB2Dialect(default.DefaultDialect):
-
     name = 'sqlalchemy_ibmi'
     max_identifier_length = 128
     encoding = 'utf-8'
@@ -676,7 +698,7 @@ class DB2Dialect(default.DefaultDialect):
     supports_empty_insert = False
 
     two_phase_transactions = False
-    savepoints =  True
+    savepoints = True
 
     statement_compiler = DB2Compiler
     ddl_compiler = DB2DDLCompiler
@@ -684,12 +706,89 @@ class DB2Dialect(default.DefaultDialect):
     preparer = DB2IdentifierPreparer
     execution_ctx_cls = DB2ExecutionContext
 
-    _reflector_cls = ibm_reflection.DB2Reflector
+    ischema = MetaData()
 
-    def __init__(self, **kw):
-        super(DB2Dialect, self).__init__(**kw)
+    sys_schemas = Table("SQLSCHEMAS", ischema,
+                        Column("TABLE_SCHEM", CoerceUnicode, key="schemaname"),
+                        schema="SYSIBM")
 
-        self._reflector = self._reflector_cls(self)
+    sys_tables = Table("SYSTABLES", ischema,
+                       Column("TABLE_SCHEMA", CoerceUnicode, key="tabschema"),
+                       Column("TABLE_NAME", CoerceUnicode, key="tabname"),
+                       Column("TABLE_TYPE", CoerceUnicode, key="tabtype"),
+                       schema="QSYS2")
+
+    sys_table_constraints = Table("SYSCST", ischema,
+                                  Column("CONSTRAINT_SCHEMA", CoerceUnicode, key="conschema"),
+                                  Column("CONSTRAINT_NAME", CoerceUnicode, key="conname"),
+                                  Column("CONSTRAINT_TYPE", CoerceUnicode, key="contype"),
+                                  Column("TABLE_SCHEMA", CoerceUnicode, key="tabschema"),
+                                  Column("TABLE_NAME", CoerceUnicode, key="tabname"),
+                                  Column("TABLE_TYPE", CoerceUnicode, key="tabtype"),
+                                  schema="QSYS2")
+
+    sys_key_constraints = Table("SYSKEYCST", ischema,
+                                Column("CONSTRAINT_SCHEMA", CoerceUnicode, key="conschema"),
+                                Column("CONSTRAINT_NAME", CoerceUnicode, key="conname"),
+                                Column("TABLE_SCHEMA", CoerceUnicode, key="tabschema"),
+                                Column("TABLE_NAME", CoerceUnicode, key="tabname"),
+                                Column("COLUMN_NAME", CoerceUnicode, key="colname"),
+                                Column("ORDINAL_POSITION", sa_types.Integer, key="colno"),
+                                schema="QSYS2")
+
+    sys_columns = Table("SYSCOLUMNS", ischema,
+                        Column("TABLE_SCHEMA", CoerceUnicode, key="tabschema"),
+                        Column("TABLE_NAME", CoerceUnicode, key="tabname"),
+                        Column("COLUMN_NAME", CoerceUnicode, key="colname"),
+                        Column("ORDINAL_POSITION", sa_types.Integer, key="colno"),
+                        Column("DATA_TYPE", CoerceUnicode, key="typename"),
+                        Column("LENGTH", sa_types.Integer, key="length"),
+                        Column("NUMERIC_SCALE", sa_types.Integer, key="scale"),
+                        Column("IS_NULLABLE", sa_types.Integer, key="nullable"),
+                        Column("COLUMN_DEFAULT", CoerceUnicode, key="defaultval"),
+                        Column("HAS_DEFAULT", CoerceUnicode, key="hasdef"),
+                        Column("IS_IDENTITY", CoerceUnicode, key="isid"),
+                        Column("IDENTITY_GENERATION", CoerceUnicode, key="idgenerate"),
+                        schema="QSYS2")
+
+    sys_indexes = Table("SYSINDEXES", ischema,
+                        Column("TABLE_SCHEMA", CoerceUnicode, key="tabschema"),
+                        Column("TABLE_NAME", CoerceUnicode, key="tabname"),
+                        Column("INDEX_SCHEMA", CoerceUnicode, key="indschema"),
+                        Column("INDEX_NAME", CoerceUnicode, key="indname"),
+                        Column("IS_UNIQUE", CoerceUnicode, key="uniquerule"),
+                        schema="QSYS2")
+
+    sys_keys = Table("SYSKEYS", ischema,
+                     Column("INDEX_SCHEMA", CoerceUnicode, key="indschema"),
+                     Column("INDEX_NAME", CoerceUnicode, key="indname"),
+                     Column("COLUMN_NAME", CoerceUnicode, key="colname"),
+                     Column("ORDINAL_POSITION", sa_types.Integer, key="colno"),
+                     Column("ORDERING", CoerceUnicode, key="ordering"),
+                     schema="QSYS2")
+
+    sys_foreignkeys = Table("SQLFOREIGNKEYS", ischema,
+                            Column("FK_NAME", CoerceUnicode, key="fkname"),
+                            Column("FKTABLE_SCHEM", CoerceUnicode, key="fktabschema"),
+                            Column("FKTABLE_NAME", CoerceUnicode, key="fktabname"),
+                            Column("FKCOLUMN_NAME", CoerceUnicode, key="fkcolname"),
+                            Column("PK_NAME", CoerceUnicode, key="pkname"),
+                            Column("PKTABLE_SCHEM", CoerceUnicode, key="pktabschema"),
+                            Column("PKTABLE_NAME", CoerceUnicode, key="pktabname"),
+                            Column("PKCOLUMN_NAME", CoerceUnicode, key="pkcolname"),
+                            Column("KEY_SEQ", sa_types.Integer, key="colno"),
+                            schema="SYSIBM")
+
+    sys_views = Table("SYSVIEWS", ischema,
+                      Column("TABLE_SCHEMA", CoerceUnicode, key="viewschema"),
+                      Column("TABLE_NAME", CoerceUnicode, key="viewname"),
+                      Column("VIEW_DEFINITION", CoerceUnicode, key="text"),
+                      schema="QSYS2")
+
+    sys_sequences = Table("SYSSEQUENCES", ischema,
+                          Column("SEQUENCE_SCHEMA", CoerceUnicode, key="seqschema"),
+                          Column("SEQUENCE_NAME", CoerceUnicode, key="seqname"),
+                          schema="QSYS2")
 
     # reflection: these all defer to an BaseDB2Reflector
     # object which selects between DB2 and AS/400 schemas
@@ -697,67 +796,224 @@ class DB2Dialect(default.DefaultDialect):
         super(DB2Dialect, self).initialize(connection)
         self.dbms_ver = getattr(connection.connection, 'dbms_ver', None)
         self.dbms_name = getattr(connection.connection, 'dbms_name', None)
-        
-    def normalize_name(self, name):
-        return self._reflector.normalize_name(name)
-
-    def denormalize_name(self, name):
-        return self._reflector.denormalize_name(name)
 
     def _get_default_schema_name(self, connection):
-        return self._reflector._get_default_schema_name(connection)
+        """Return: current setting of the schema attribute"""
+        default_schema_name = connection.execute(
+            u'SELECT CURRENT_SCHEMA FROM SYSIBM.SYSDUMMY1').scalar()
+        if isinstance(default_schema_name, str):
+            default_schema_name = default_schema_name.strip()
+        return self.normalize_name(default_schema_name)
+
+    @property
+    def default_schema_name(self):
+        return self.default_schema_name
 
     def has_table(self, connection, table_name, schema=None):
-        return self._reflector.has_table(connection, table_name, schema=schema)
+        current_schema = self.denormalize_name(
+            schema or self.default_schema_name)
+        table_name = self.denormalize_name(table_name)
+        if current_schema:
+            whereclause = sql.and_(
+                self.sys_tables.c.tabschema == current_schema,
+                self.sys_tables.c.tabname == table_name)
+        else:
+            whereclause = self.sys_tables.c.tabname == table_name
+        s = sql.select([self.sys_tables], whereclause)
+        c = connection.execute(s)
+        return c.first() is not None
 
     def has_sequence(self, connection, sequence_name, schema=None):
-        return self._reflector.has_sequence(connection, sequence_name,
-                        schema=schema)
+        current_schema = self.denormalize_name(
+            schema or self.default_schema_name)
+        sequence_name = self.denormalize_name(sequence_name)
+        if current_schema:
+            whereclause = sql.and_(
+                self.sys_sequences.c.seqschema == current_schema,
+                self.sys_sequences.c.seqname == sequence_name)
+        else:
+            whereclause = self.sys_sequences.c.seqname == sequence_name
+        s = sql.select([self.sys_sequences.c.seqname], whereclause)
+        c = connection.execute(s)
+        return c.first() is not None
 
+    @reflection.cache
     def get_schema_names(self, connection, **kw):
-        return self._reflector.get_schema_names(connection, **kw)
+        sysschema = self.sys_schemas
+        query = sql.select([sysschema.c.schemaname],
+                           sql.not_(sysschema.c.schemaname.like('SYS%')),
+                           sql.not_(sysschema.c.schemaname.like('Q%')),
+                           order_by=[sysschema.c.schemaname]
+                           )
+        return [self.normalize_name(r[0]) for r in connection.execute(query)]
 
-
+    # Retrieves a list of table names for a given schema
+    @reflection.cache
     def get_table_names(self, connection, schema=None, **kw):
-        return self._reflector.get_table_names(connection, schema=schema, **kw)
+        current_schema = self.denormalize_name(schema or self.default_schema_name)
+        systbl = self.sys_tables
+        query = sql.select([systbl.c.tabname]). \
+            where(systbl.c.tabtype == 'T'). \
+            where(systbl.c.tabschema == current_schema). \
+            order_by(systbl.c.tabname)
+        return [self.normalize_name(r[0]) for r in connection.execute(query)]
 
+    @reflection.cache
     def get_view_names(self, connection, schema=None, **kw):
-        return self._reflector.get_view_names(connection, schema=schema, **kw)
+        current_schema = self.denormalize_name(
+            schema or self.default_schema_name)
 
+        query = sql.select([self.sys_views.c.viewname],
+                           self.sys_views.c.viewschema == current_schema,
+                           order_by=[self.sys_views.c.viewname]
+                           )
+        return [self.normalize_name(r[0]) for r in connection.execute(query)]
+
+    @reflection.cache
     def get_view_definition(self, connection, viewname, schema=None, **kw):
-        return self._reflector.get_view_definition(
-                                connection, viewname, schema=schema, **kw)
+        current_schema = self.denormalize_name(
+            schema or self.default_schema_name)
+        viewname = self.denormalize_name(viewname)
 
+        query = sql.select([self.sys_views.c.text],
+                           self.sys_views.c.viewschema == current_schema,
+                           self.sys_views.c.viewname == viewname
+                           )
+        return connection.execute(query).scalar()
+
+    @reflection.cache
     def get_columns(self, connection, table_name, schema=None, **kw):
-        return self._reflector.get_columns(
-                                connection, table_name, schema=schema, **kw)
+        current_schema = self.denormalize_name(schema or self.default_schema_name)
+        table_name = self.denormalize_name(table_name)
+        syscols = self.sys_columns
 
+        query = sql.select([syscols.c.colname,
+                            syscols.c.typename,
+                            syscols.c.defaultval, syscols.c.nullable,
+                            syscols.c.length, syscols.c.scale,
+                            syscols.c.isid, syscols.c.idgenerate],
+                           sql.and_(
+                               syscols.c.tabschema == current_schema,
+                               syscols.c.tabname == table_name
+                           ),
+                           order_by=[syscols.c.colno]
+                           )
+        sa_columns = []
+        for r in connection.execute(query):
+            coltype = r[1].upper()
+            if coltype in ['DECIMAL', 'NUMERIC']:
+                coltype = self.ischema_names.get(coltype)(int(r[4]), int(r[5]))
+            elif coltype in ['CHARACTER', 'CHAR', 'VARCHAR',
+                             'GRAPHIC', 'VARGRAPHIC']:
+                coltype = self.ischema_names.get(coltype)(int(r[4]))
+            else:
+                try:
+                    coltype = self.ischema_names[coltype]
+                except KeyError:
+                    util.warn("Did not recognize type '%s' of column '%s'" %
+                              (coltype, r[0]))
+                    coltype = coltype = sa_types.NULLTYPE
+
+            sa_columns.append({
+                'name': self.normalize_name(r[0]),
+                'type': coltype,
+                'nullable': r[3] == 'Y',
+                'default': r[2],
+                'autoincrement': (r[6] == 'YES') and (r[7] != None),
+            })
+        return sa_columns
+
+    @reflection.cache
     def get_primary_keys(self, connection, table_name, schema=None, **kw):
-        return self._reflector.get_primary_keys(
-                                connection, table_name, schema=schema, **kw)
+        current_schema = self.denormalize_name(
+            schema or self.default_schema_name)
+        table_name = self.denormalize_name(table_name)
+        sysconst = self.sys_table_constraints
+        syskeyconst = self.sys_key_constraints
 
+        query = sql.select([syskeyconst.c.colname, sysconst.c.tabname],
+                           sql.and_(
+                               syskeyconst.c.conschema == sysconst.c.conschema,
+                               syskeyconst.c.conname == sysconst.c.conname,
+                               sysconst.c.tabschema == current_schema,
+                               sysconst.c.tabname == table_name,
+                               sysconst.c.contype == 'PRIMARY KEY'
+                           ), order_by=[syskeyconst.c.colno])
+
+        return [self.normalize_name(key[0])
+                for key in connection.execute(query)]
+
+    @reflection.cache
     def get_foreign_keys(self, connection, table_name, schema=None, **kw):
-        return self._reflector.get_foreign_keys(
-                                connection, table_name, schema=schema, **kw)
-        
-    def get_incoming_foreign_keys(self, connection, table_name, schema=None, **kw):
-        return self._reflector.get_incoming_foreign_keys(
-                                connection, table_name, schema=schema, **kw)
+        default_schema = self.default_schema_name
+        current_schema = self.denormalize_name(schema or default_schema)
+        default_schema = self.normalize_name(default_schema)
+        table_name = self.denormalize_name(table_name)
+        sysfkeys = self.sys_foreignkeys
+        query = sql.select([sysfkeys.c.fkname, sysfkeys.c.fktabschema, \
+                            sysfkeys.c.fktabname, sysfkeys.c.fkcolname, \
+                            sysfkeys.c.pkname, sysfkeys.c.pktabschema, \
+                            sysfkeys.c.pktabname, sysfkeys.c.pkcolname],
+                           sql.and_(
+                               sysfkeys.c.fktabschema == current_schema,
+                               sysfkeys.c.fktabname == table_name
+                           ),
+                           order_by=[sysfkeys.c.colno]
+                           )
+        fschema = {}
+        for r in connection.execute(query):
+            if r[0] not in fschema:
+                referred_schema = self.normalize_name(r[5])
 
+                # if no schema specified and referred schema here is the
+                # default, then set to None
+                if schema is None and \
+                        referred_schema == default_schema:
+                    referred_schema = None
+
+                fschema[r[0]] = {'name': self.normalize_name(r[0]),
+                                 'constrained_columns': [self.normalize_name(r[3])],
+                                 'referred_schema': referred_schema,
+                                 'referred_table': self.normalize_name(r[6]),
+                                 'referred_columns': [self.normalize_name(r[7])]}
+            else:
+                fschema[r[0]]['constrained_columns'].append(self.normalize_name(r[3]))
+                fschema[r[0]]['referred_columns'].append(self.normalize_name(r[7]))
+        return [value for key, value in fschema.items()]
+
+    # Retrieves a list of index names for a given schema
+    @reflection.cache
     def get_indexes(self, connection, table_name, schema=None, **kw):
-        return self._reflector.get_indexes(
-                                connection, table_name, schema=schema, **kw)
-        
+        current_schema = self.denormalize_name(
+            schema or self.default_schema_name)
+        table_name = self.denormalize_name(table_name)
+        sysidx = self.sys_indexes
+        syskey = self.sys_keys
+
+        query = sql.select([sysidx.c.indname,
+                            sysidx.c.uniquerule, syskey.c.colname], sql.and_(
+            syskey.c.indschema == sysidx.c.indschema,
+            syskey.c.indname == sysidx.c.indname,
+            sysidx.c.tabschema == current_schema,
+            sysidx.c.tabname == table_name
+        ), order_by=[syskey.c.indname, syskey.c.colno]
+                           )
+        indexes = {}
+        for r in connection.execute(query):
+            key = r[0].upper()
+            if key in indexes:
+                indexes[key]['column_names'].append(self.normalize_name(r[2]))
+            else:
+                indexes[key] = {
+                    'name': self.normalize_name(r[0]),
+                    'column_names': [self.normalize_name(r[2])],
+                    'unique': r[1] == 'Y'
+                }
+        return [value for key, value in indexes.items()]
+
+    @reflection.cache
     def get_unique_constraints(self, connection, table_name, schema=None, **kw):
-        return self._reflector.get_unique_constraints(
-                                connection, table_name, schema=schema, **kw)
+        uniqueConsts = []
+        return uniqueConsts
 
 
-# legacy naming
-IBM_DBCompiler = DB2Compiler
-IBM_DBDDLCompiler = DB2DDLCompiler
-IBM_DBIdentifierPreparer = DB2IdentifierPreparer
-IBM_DBExecutionContext = DB2ExecutionContext
-IBM_DBDialect = DB2Dialect
-
-dialect = DB2Dialect
